continuous run now implemented, _highscore begins as 5 and is continually incremented when _score > _highscore. Whenever a command times out, _score is reset to 0 and _highscore remains at whatever the longest run was. The incrementing of the progress bar also works. Done for this week

+++ Glossalalia/Glossalalia/GALevelViewController.h
@@ -29,8 +29,8 @@
 @property NSObject <GALevelViewControllerDelegate> *delegate;
 @property GADataHandler *dataHandler;
 
-@property int score;
-@property int highScore;
+@property float score;
+@property float highScore;
 @property int numWordsCorrect;
 @property int numButtonWordsPerPlayer;
 @property CGFloat commandCompletionTimeLimit;
